add edition field to rustfmt.toml file
This is important. Without this line, any invocation of `rustfmt`
directly is going to read rustfmt.toml, but fall back to the default
edition, which is 2015 (or was until recently). This causes the code
to be formatted slightly differently from someone who runs `cargo
fmt`. The latter looks up the edition from the Cargo.toml file.

In Emacs for example the standard way to use rustfmt is via the binary
directly and *not* via Cargo. So any editor / tooling that works like
Emacs will produce wrongly formatted code despite the fact that
rustfmt is used.