simplify `poseidon_safe_domain_separator` (#25)
The doc comment previously explained that the function was specific to
a single Poseidon permutation. However, it still tried to be
generic. As a result the code was more complex than necessary (3
additional generic arguments) and technically broken for any
permutation that does not use 24 field elements.

Instead it now is indeed specific to the poseidon2 permutation with 24
field elements.

For the use case of producing packed data, we simply broadcast the
result after calling the function. This is only used in the preamble
of the `compute_tree_leaves` code anyway.