avoid heap allocations in `compute_tree_leaves` by using `for_each_init`
Each rayon worker job had to allocate the full `packed_leaf_input`. We
now use `for_each_init` to preallocate a vector for every Rayon worker
instead. We overwrite the entire vector in every job, so not even a
need to `fill(0)` the vector in each job.

This drops another ~100k allocations when running the 2^32 bench over
30s.

Brings us down to only 3k temporary allocations total in that time frame.