avoid heap allocations in 2/3 branches of `apply`
No need for a `Vec` in these two branches as we know at compile time
how much data is required for each input.

Only relevant if `apply` is part of a hot code path, which normally is
unlikely to be the case. Still, the code is not significantly more,
only more ugly :(

It gets rid of a large number of allocations when running the 2^8
benchmark case.